[#2] Chore: 스토리보드 제거
