+++ githubIssueManager/githubIssueManager/ViewController.swift
@@ -1,19 +1,9 @@
-//
-//  ViewController.swift
-//  githubIssueManager
-//
-//  Created by 이준우 on 2022/06/17.
-//
-
 import UIKit
 
 class ViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
-        // Do any additional setup after loading the view.
     }
-
-
 }
